test: move tests into several files
